src: non-canonical cast (double) is deprecated

+++ src/opnsense/scripts/interfaces/traffic_stats.php
@@ -33,7 +33,7 @@ function map_ifs($ifs, $data)
 {
     $result = ["interfaces" => []];
     $temp = gettimeofday();
-    $result['time'] = (double)$temp["sec"] + (double)$temp["usec"] / 1000000.0;
+    $result['time'] = (float)$temp['sec'] + (float)$temp['usec'] / 1000000.0;
 
     foreach ($ifs as $interfaceKey => $itf) {
         if (array_key_exists($itf['if'], $data)) {
